Refactoring Namespace
