+++ sample/config/example_backup_storage.php
