+++ sample/config/example_db_codeigniter.php
