+++ src/Model/BaseModel.php
@@ -16,8 +16,8 @@ use Illuminate\Container\Container;
 use nguyenanhung\MyDebug\Debug;
 use nguyenanhung\MyDatabase\ProjectInterface;
 use nguyenanhung\MyDatabase\ModelInterface;
-use nguyenanhung\MyDatabase\Version;
-use nguyenanhung\MyDatabase\Helper;
+use nguyenanhung\MyDatabase\Traits\Common;
+use nguyenanhung\MyDatabase\Traits\Version;
 
 /**
  * Class BaseModel
@@ -38,7 +38,7 @@ use nguyenanhung\MyDatabase\Helper;
  */
 class BaseModel implements ProjectInterface, ModelInterface, BaseModelInterface
 {
-    use Version, Helper;
+    use Version, Common;
 
     /** @var object Đối tượng khởi tạo dùng gọi đến Class Debug \nguyenanhung\MyDebug\Debug */
     protected $debug;
