+++ src/Model/PDOUtilsModel.php
@@ -15,8 +15,8 @@ use FaaPz\PDO\Database;
 use nguyenanhung\MyDebug\Debug;
 use nguyenanhung\MyDatabase\ProjectInterface;
 use nguyenanhung\MyDatabase\ModelInterface;
-use nguyenanhung\MyDatabase\Version;
-use nguyenanhung\MyDatabase\Helper;
+use nguyenanhung\MyDatabase\Traits\Common;
+use nguyenanhung\MyDatabase\Traits\Version;
 
 /**
  * Class PDOUtilsModel
@@ -27,7 +27,7 @@ use nguyenanhung\MyDatabase\Helper;
  */
 class PDOUtilsModel implements ProjectInterface, ModelInterface, PDOUtilsModelInterface
 {
-    use Version, Helper;
+    use Version, Common;
 
     /** @var object Đối tượng khởi tạo dùng gọi đến Class Debug \nguyenanhung\MyDebug\Debug */
     protected $debug;
